+++ src/game/scripts/vscripts/commands.lua
@@ -1,4 +1,5 @@
 Commands = Commands or class({})
+local Timers = require('easytimers')
 
 function Commands:CheckArgs( args, toCheck )
     for k,v in pairs(args) do
@@ -42,54 +43,150 @@ function Commands:OnPlayerChat(keys)
     ----------------------------
     -- Vote Commands
     ----------------------------
+    -- If a vote fails, players cannot call another vote for 5 minutes, to prevent abuse.
+    if ingame.votesBlocked[playerID] == true then 
+        util:DisplayError(0, "#votingCooldown")
+        return
+    end
+    local voteCooldown = 300
+
     if string.find(command, "-antirat") or string.find(command, "-ar")then
         if OptionManager:GetOption('antiRat') == 0 and not ingame.voteEnabledCheatMode then
+            
+            -- If a vote has been called of this type recently, block 
+            if ingame.votesBlocked["antirat"] == true then
+                util:DisplayError(0, "#voteCooldown")
+                return
+            end
+
+            ingame.votesBlocked["antirat"] = true
+            -- Temporarily block future votes if the vote is not succesful
+            ingame.votesBlocked[playerID] = true          
+            
+            Timers:CreateTimer( function()
+                ingame.votesBlocked["antirat"] = false
+                ingame.votesBlocked[playerID] = false
+            end, DoUniqueString('voteTempBlock'), voteCooldown)
+
             -- In all_allowed map, votes needed is only 50% of players (rounded up)
             local percentNeeded = 100
             if GetMapName() == 'all_allowed' then
                 percentNeeded = 50
             end
+          
             util:CreateVoting("lodVotingAntirat", playerID, 20, percentNeeded, function()
+                ingame.votesBlocked[playerID] = false
                 OptionManager:SetOption('antiRat', 1) 
                 ingame:giveAntiRatProtection()
                 ingame.voteAntiRat = true
                 EmitGlobalSound("Event.CheatEnabled")
                 GameRules:SendCustomMessage('Enough players voted to enable anti-rat protection. <font color=\'#70EA72\'>Tier 3 towers cannot be destroyed until all other towers are gone.</font>.',0,0)
             end)
+        elseif OptionManager:GetOption('antiRat') == 1 then
+            util:DisplayError(0, "#antiRatAlreadyOn")
         end
     elseif string.find(command, "-doublecreeps") or string.find(command, "-dc") then
-       if not ingame.voteDoubleCreeps then 
+       if not ingame.voteDoubleCreeps and OptionManager:GetOption('neutralMultiply') < 2 then 
+            
+            if ingame.votesBlocked["doublecreeps"] == true then
+                util:DisplayError(0, "#voteCooldown")
+                return
+            end
+
+            ingame.votesBlocked["doublecreeps"] = true
+            ingame.votesBlocked[playerID] = true 
+
+            Timers:CreateTimer( function()
+                ingame.votesBlocked[playerID] = false
+                ingame.votesBlocked["doublecreeps"] = false
+            end, DoUniqueString('voteTempBlock'), voteCooldown)
+
             local percentNeeded = 100
             if GetMapName() == 'all_allowed' then
                 percentNeeded = 50
             end
+
             util:CreateVoting("lodVotingDoubleCreeps", playerID, 20, percentNeeded, function()
+                ingame.votesBlocked[playerID] = false
                 OptionManager:SetOption('neutralMultiply', 2)
                -- pregame:multiplyNeutrals()
                 ingame.voteDoubleCreeps = true
                 EmitGlobalSound("Event.CheatEnabled")
                 GameRules:SendCustomMessage('Enough players voted to enable double neutrals. <font color=\'#70EA72\'>Neutral creep camps are now doubled</font>.',0,0)
             end)
+        elseif OptionManager:GetOption('neutralMultiply') > 1 then
+            util:DisplayError(0, "#multiplyAlreadyOn")
         end
     elseif string.find(command, "-enablecheat") or string.find(command, "-ec") then
-        if not ingame.voteEnabledCheatMode then       
+        if not ingame.voteEnabledCheatMode and not Convars:GetBool("sv_cheats") then
+            -- Check if this particular vote has been called recently, if so ignore it
+            if ingame.votesBlocked["enablecheat"] == true then
+                util:DisplayError(0, "#voteCooldown")
+                return
+            end
+
+            -- Block starting votes for the player and the particular vote, temporarily
+            ingame.votesBlocked["enablecheat"] = true
+            ingame.votesBlocked[playerID] = true         
+            
+            --Re-allow the player to vote, and the particlar vote to be called, after a timer
+            Timers:CreateTimer( function()
+                ingame.votesBlocked["enablecheat"] = false
+                ingame.votesBlocked[playerID] = false   
+            end, DoUniqueString('voteTempBlock'), voteCooldown)
+
             util:CreateVoting("lodVotingEnableCheatMode", playerID, 20, 100, function()
+                ingame.votesBlocked[playerID] = false
                 ingame.voteEnabledCheatMode = true
                 EmitGlobalSound("Event.CheatEnabled")
                 GameRules:SendCustomMessage('<font color=\'#70EA72\'>Everbody voted to enable cheat mode. Cheat mode enabled</font>.',0,0)
             end)
+        elseif ingame.voteEnabledCheatMode or Convars:GetBool("sv_cheats") then
+            util:DisplayError(0, "#cheatModeAlreadyOn")
         end
     elseif string.find(command, "-enablekamikaze") or string.find(command, "-ek") then
         if not ingame.voteDisableAntiKamikaze then
+
+            if ingame.votesBlocked["enablekamikaze"] == true then
+                util:DisplayError(0, "#voteCooldown")
+                return
+            end
+
+            ingame.votesBlocked["enablekamikaze"] = true
+            ingame.votesBlocked[playerID] = true    
+
+            Timers:CreateTimer( function()
+                ingame.votesBlocked["enablekamikaze"] = false
+                ingame.votesBlocked[playerID] = false
+            end, DoUniqueString('voteTempBlock'), voteCooldown)
+
             util:CreateVoting("lodVotingEnableKamikaze", playerID, 20, 100, function()
+                ingame.votesBlocked[playerID] = false 
                 ingame.voteDisableAntiKamikaze = true
                 EmitGlobalSound("Event.CheatEnabled")
                 GameRules:SendCustomMessage('Everbody voted to disable the anti-Kamikaze mechanic. <font color=\'#70EA72\'>No more peanlty for dying 3 times within 60 seconds</font>.',0,0)
             end)
+        elseif ingame.voteDisableAntiKamikaze then
+            util:DisplayError(0, "#kamikazeAlreadyDeactivated")
         end
     elseif string.find(command, "-enablebuilder") or string.find(command, "-eb") and OptionManager:GetOption('allowIngameHeroBuilder') == false then
-        if not ingame.voteEnableBuilder then
+        if not ingame.voteEnableBuilder and OptionManager:GetOption('allowIngameHeroBuilder') ~= true then
+            
+            if ingame.votesBlocked["enablebuilder"] == true then
+                util:DisplayError(0, "#voteCooldown")
+                return
+            end
+
+            ingame.votesBlocked["enablebuilder"] = true
+            ingame.votesBlocked[playerID] = true          
+            
+            Timers:CreateTimer( function()
+                ingame.votesBlocked[playerID] = false
+                ingame.votesBlocked["enablebuilder"] = false
+            end, DoUniqueString('voteTempBlock'), voteCooldown)
+
             util:CreateVoting("lodVotingEnableHeroBuilder", playerID, 20, 100, function()
+                ingame.votesBlocked[playerID] = false
                 network:enableIngameHeroEditor()
                 OptionManager:SetOption('allowIngameHeroBuilder', 1)
                 if util:GetActivePlayerCountForTeam(DOTA_TEAM_GOODGUYS) > 0 and util:GetActivePlayerCountForTeam(DOTA_TEAM_GOODGUYS) > 0 then
@@ -99,10 +196,27 @@ function Commands:OnPlayerChat(keys)
                 EmitGlobalSound("Event.CheatEnabled")
                 GameRules:SendCustomMessage('Everbody voted to enable the ingame hero builder. <font color=\'#70EA72\'>You can now change your hero build mid-game</font>.',0,0)
             end)
+        elseif OptionManager:GetOption('allowIngameHeroBuilder') == true or ingame.voteEnableBuilder then
+            util:DisplayError(0, "#heroBuilderAlreadyOn")
         end
     elseif string.find(command, "-enablerespawn") or string.find(command, "-er") then
         if not ingame.voteDisableRespawnLimit then
+            
+            if ingame.votesBlocked["enablerespawn"] == true then
+                util:DisplayError(0, "#voteCooldown")
+                return
+            end
+
+            ingame.votesBlocked["enablerespawn"] = true
+            ingame.votesBlocked[playerID] = true          
+            
+            Timers:CreateTimer( function()
+                ingame.votesBlocked["enablerespawn"] = false
+                ingame.votesBlocked[playerID] = false
+            end, DoUniqueString('voteTempBlock'), voteCooldown)
+
             util:CreateVoting("lodVotingEnableRespawn", playerID, 20, 100, function()
+                ingame.votesBlocked[playerID] = false
                 ingame.voteDisableRespawnLimit = true
                 if ingame.origianlRespawnRate ~= nil then
                     OptionManager:SetOption('respawnModifierPercentage', ingame.origianlRespawnRate)
@@ -110,11 +224,28 @@ function Commands:OnPlayerChat(keys)
                 EmitGlobalSound("Event.CheatEnabled")
                 GameRules:SendCustomMessage('Everbody voted to disable the increasing-spawn-rate mechanic. <font color=\'#70EA72\'>Respawn rates no longer increase after 40 minutes</font>. Respawn rate is now '.. OptionManager:GetOption('respawnModifierPercentage') .. '%.',0,0)
             end)
+        elseif ingame.voteDisableRespawnLimit then
+            util:DisplayError(0, "#respawnAlreadyDeactivated")
         end
     elseif string.find(command, "-switchteam") then
         local team = PlayerResource:GetTeam(playerID)
         if (ingame.needsTeamBalance and ingame.takeFromTeam == team) or util:isSinglePlayerMode() or IsInToolsMode() then
+            
+            if ingame.votesBlocked["switchteam"] == true then
+                util:DisplayError(0, "#voteCooldown")
+                return
+            end
+
+            ingame.votesBlocked["switchteam"] = true          
+            ingame.votesBlocked[playerID] = true          
+            
+            Timers:CreateTimer( function()
+                ingame.votesBlocked["switchteam"] = false
+                ingame.votesBlocked[playerID] = false
+            end, DoUniqueString('voteTempBlock'), voteCooldown)
+
             util:CreateVoting("lodVotingSwitchTeam", playerID, 20, 100, function()
+                ingame.votesBlocked[playerID] = false
                 local oldTeam = PlayerResource:GetCustomTeamAssignment(playerID)
                 local newTeam = otherTeam(oldTeam)
                 local uMoney = PlayerResource:GetUnreliableGold(playerID)